+++ src/nti/externalization/internalization/fields.py
@@ -21,7 +21,6 @@ from six import reraise
 from zope.interface import implementedBy
 
 from zope.schema.interfaces import IField
-from zope.schema.interfaces import IFromUnicode
 from zope.schema.interfaces import SchemaNotProvided
 from zope.schema.interfaces import SchemaNotCorrectlyImplemented
 from zope.schema.interfaces import ValidationError
@@ -35,7 +34,6 @@ from zope.schema.fieldproperty import FieldUpdatedEvent
 from zope.event import notify
 
 IField_providedBy = IField.providedBy
-IFromUnicode_providedBy = IFromUnicode.providedBy
 
 __all__ = [
     'validate_field_value',
@@ -144,22 +142,21 @@ def _adapt_sequence(field, value):
     # using it. Some other things do not, for example nti.schema.field.Variant
     # They might provide a `fromObject` function to do the conversion
     # The field may be able to handle the whole thing by itself or we may need
-    # to do the individual objects
+    # to do the individual objects. If there was a `fromObject` for the field,
+    # we called it already.
 
     # The conversion process may raise TypeError
-    if hasattr(field, 'fromObject'):
-        value = field.fromObject(value)
+    value_type = field.value_type
+    if hasattr(value_type, 'fromObject'):
+        converter = value_type.fromObject
+    elif hasattr(value_type, 'schema'):
+        converter = value_type.schema
     else:
-        if hasattr(field.value_type, 'fromObject'):
-            converter = field.value_type.fromObject
-        elif hasattr(field.value_type, 'schema'):
-            converter = field.value_type.schema
-        else:
-            raise CannotConvertSequenceError(
-                "Don't know how to convert sequence %r for field %s"
-                % (value, field))
+        raise CannotConvertSequenceError(
+            "Don't know how to convert sequence %r for field %s"
+            % (value, field))
 
-        value = [converter(v) for v in value]
+    value = [converter(v) for v in value]
 
     return value
 
@@ -263,6 +260,12 @@ def _handle_WrongContainedType(field_name, field, value):
 
     return value
 
+_CONVERTERS = (
+    ('fromUnicode', text_type),
+    ('fromBytes', bytes),
+    ('fromObject', object)
+)
+
 def validate_field_value(self, field_name, field, value):
     """
     Given a :class:`zope.schema.interfaces.IField` object from a schema
@@ -282,10 +285,16 @@ def validate_field_value(self, field_name, field, value):
     """
     field = field.bind(self)
     try:
-        if isinstance(value, text_type) and IFromUnicode_providedBy(field):
-            value = field.fromUnicode(value)  # implies validation
+        for meth_name_kind in _CONVERTERS:
+            if isinstance(value, meth_name_kind[1]):
+                meth = getattr(field, meth_name_kind[0], None)
+                if meth is not None:
+                    value = meth(value)
+                    break
         else:
+            # Here if we do not break out of the loop.
             field.validate(value)
+
     except SchemaNotProvided:
         # Raised by Object fields
         value = _handle_SchemaNotProvided(field_name, field, value)